readme.md
The “Grocery Management System” is developed to facilitate the work of the grocery store keeper and the customer who buys the products. This system consists of the methods which can relate the perpetual crisis while managing the rhetoric quandaries concerning the authoritarian source .The proposed system will avail the customer to opt for the type of grocery items they require and forward the list to the admin then after the admin suggests a team and pack the items followed by the distribution or delivery of the desired products to the doorstep of the customer. The whole system of grocery management is controlled and sustained by an Admin, who holds the authority to terminate any account and change any information within the system if required. This entity gains momentum while keeping intact the system to run smoothly and error-free.

The project also helps to maintain the supply chain. This is achieved by knowing about the retailer, distributor, and customer associated with a product. This feature will help reduce managerial problems faced in the supply chain. This will help us reduce the spoilage of technology and resources available with us thus helping the Grocery owner gain maximum profits with the finest technology hence depicting how technology can be used in any scenario possible hence solving every problem around us.